refactor(api): #3803 refacto interface port
