+++ packages/api/src/interfaces/http/DepositScdlProcess.http.ts
@@ -19,7 +19,7 @@ import {
     FormField,
 } from "tsoa";
 import DepositScdlLogDtoMapper from "../../modules/deposit-scdl-process/deposit-scdl-log.dto.mapper";
-import { depositScdlProcessService } from "../../configurations/di-container";
+import { depositScdlProcessService } from "../../init-services";
 
 @Route("/parcours-depot")
 @Security("jwt")
